use domain/relativePath for iOS, replacing domain-relativePath
iOS filesystem backups use domain-relativePath for hashing purposes, but
domains can also contain dashes, so there isn't a simple way to separate
domain and relative path. This is important if we want to use
Manifest.db to find files, as these two components are in separate
tables. Switch from '-' to '/' as the separator to disambiguate.