Switch back from per-device event results to a single chronological list
It seems that it's easier for the GUI if we use a single list. I expect
it will be easier for API users as well; if you want a per-device result
set, it's easy enough to perform multiple single-device graphql calls.

The per-device lists were made in the hope that there would be a nice
way to populate columns in the GUI without preprocessing, but that
doesn't seem to be the case given that the timestamps need to line up
across devices.