Add 'source' attribute for all Events
The idea behind 'source' is that it is something fairly close to a
filesystem directory or filename, which API users can use to filter
results. It's motivated by the generic image finder, which turns up a
lot of app default media (WhatsApp wallpapers and so on). All this is in
a specific directory, but it's not great to automatically hide it.
Allowing filtering by (effectively) directory allows users to see and
then discard the data.