Replace 'source' with 'category' for generic events and display it
Generic events (which don't belong to a specific app) now include a
genericEventInfo object which contains a 'category' field. This replaces
'source' introduced in previous commits and is supposed to indicate
"where the event came from" in the absence of a specific app source. For
the Android generic media provider, the category is the filesystem
directory which contains the media.

The point is to allow filtering by category.

These are now displayed in the UI.

This also fixes a bug relating in which media events would not include
provider information.