Add Event.sender: Contact
Add the 'sender' field, a Contact, to Event as a mandatory field. This
means we can no longer have a sender of None, which in turn means we
need special "contacts" for events which were sent by the device
operator or could not be said to be generated explicitly by a person.
The advantage of doing so is that we start moving closer to being able
to filter operator- or system-generated events.

- make 'sender' (a Contact) part of Event and make it mandatory
- Add special contacts representing the device operator, the device
itself, and all apps on the device
- Move provider names and friendly names to central location so generic
provider can use them
- Add "is user generated" flag for generic media
- Use generic names in all providers and ensure we always provide
'sender'